Refactor: move media models to media/models
