+++ app/src/main/java/fr/free/nrw/commons/media/MediaDetailFragment.java
@@ -12,7 +12,7 @@ import static fr.free.nrw.commons.description.EditDescriptionConstants.WIKITEXT;
 import static fr.free.nrw.commons.upload.mediaDetails.UploadMediaDetailFragment.LAST_LOCATION;
 import static fr.free.nrw.commons.utils.LangCodeUtils.getLocalizedResources;
 import android.annotation.SuppressLint;
-import fr.free.nrw.commons.data.models.media.Caption;
+import fr.free.nrw.commons.media.media.Caption;
 import android.app.AlertDialog;
 import android.content.Context;
 import android.content.Intent;
@@ -74,7 +74,7 @@ import fr.free.nrw.commons.category.CategoryEditSearchRecyclerViewAdapter;
 import fr.free.nrw.commons.category.CategoryEditSearchRecyclerViewAdapter.Callback;
 import fr.free.nrw.commons.contributions.ContributionsFragment;
 import fr.free.nrw.commons.coordinates.CoordinateEditHelper;
-import fr.free.nrw.commons.data.models.media.IdAndCaptions;
+import fr.free.nrw.commons.media.media.IdAndCaptions;
 import fr.free.nrw.commons.delete.DeleteHelper;
 import fr.free.nrw.commons.delete.ReasonBuilder;
 import fr.free.nrw.commons.description.DescriptionEditActivity;
